requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
-e git+https://github.com/django/django.git@b435f82939edf70674856e0e1cd63973c2e0a1d1#egg=Django
django-crispy-forms==1.7.2
django-six==1.0.4
idna==2.8
oauthlib==3.0.1
Pillow==5.4.1
pkg-resources==0.0.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.2.4
urllib3==1.24.1
whitenoise==4.1.2
gunicorn==19.9.0
django-heroku==0.3.1
dj-database-url==0.5.0
psycopg2==2.7.7